pyproject.toml
[project]
name = "network-security"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.6.15",
    "dill>=0.4.0",
    "jupyter>=1.1.1",
    "lux-api>=0.5.1",
    "numpy>=2.1.3",
    "pandas>=2.3.1",
    "pip>=25.1.1",
    "pyaml>=25.5.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "ydata-profiling>=4.16.1",
]
